Fix `downloadButton()` not rendering in rmarkdown documents
Fixes https://github.com/rstudio/shiny/issues/2392

The downloadButton has a very simple longstanding issue where it does not return the tag explicitly. This is only an issue depending on how the function is called.
For example: `tag <- downloadButton(...)` has always worked fine, as has `anotherFunction(downloadButton(...))`.
It is likely this issue does not show in the normal shiny app because it uses one of these methods.
However `downloadButton(...)` didn't not work correctly, which means rmarkdown documents didn't show the downloadButton correctly.