make mathjax configurable
Provide two option, `MathJax_url` and `MathJax_config`, to enable the modification of the default mathjax.js.
Since the response of `mathjax.rstudio.com` could be slower in some regions, it is better to provide a default value and a configure technique to enable the modification of the default mathjax url.

change the names of the two newly added options

update the doc of newly added shiny.mathjax.*