insertui: use 'beforeEnd' by default without manually setting it
'beforeEnd' is a good default, it's what I've always used as well. However, by looking at the function's signature, because the vector of options starts with "beforeBegin", I always assumed that "beforeBegin" was the default. I do realize now that it's not true and that the documentation for `where` does say this, but I think it will be even more clear -- both documentation-wise and code-wise -- if we just have 'beforeEnd' naturally be the default