Update modal.R to not hardcode h4
This ensures that the modal title is not hard-coded to be an h4. Instead, it will be a div with an appropriate id that can be used to set an aria-labelledby attribute on the modal container parent. It also only applies the attribute when the title is specified, as a title is optional. This leaves users to specify the title as whatever kind of element they want (or plain text).